pyproject.toml
[tool.poetry]
name = "x-backend"
version = "0.1.0"
description = "X Clone API"
authors = ["Rafael Corrêa <rafaelscdev@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.9,<4.0"
django = "^4.2.10"
djangorestframework = "^3.14.0"
djangorestframework-simplejwt = "^5.3.1"
django-cors-headers = "^4.3.1"
psycopg2-binary = "^2.9.9"
python-dotenv = "^1.0.1"
Pillow = "^11.2.1"
django-extensions = "^3.2.3"
whitenoise = "^6.6.0"
drf-nested-routers = "^0.93.5"
gunicorn = "^21.2.0"
dj-database-url = "^2.1.0"
waitress = "^3.0.2"
pytest = "^8.3.5"
factory-boy = "^3.3.3"

[tool.poetry.group.dev.dependencies]
flake8 = "^7.0.0"
black = "^24.2.0"
isort = "^5.13.2"
pytest-django = "^4.11.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py39']
include = '\.pyi?$'

[tool.isort]
profile = "black"
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
line_length = 88

[flake8]
max-line-length = 88
extend-ignore = ["E203", "W503"]
exclude = [
    ".git",
    "__pycache__",
    "build",
    "dist",
    "migrations"
]